Moved NotesAdapter and NoteViewHolder to the adapter package
